requirements.txt
######## MLFLOW / SQL data packages
mlflow==2.2.2
catboost>=0.26
boto3==1.17.90
python-dotenv==0.17.1

############ Deep learning packages
# tensorflow
scikit-learn

# ######### Image processing  packages
# # opencv-python
# # opencv-python-headless==4.5.3.56
# scikit-image

# ######### Visualization packages
# seaborn
# plotly 

# # ######## Web API packages
# fastapi 
# uvicorn
# Jinja2
# python-multipart

# ##### MySQL packages
# mysql-connector-python  # MySQL
# psycopg2 # PostgreSQLs

##### development
jupyterlab
pytest
flake8
black

##### General packages
matplotlib
numpy
pandas
python-box==6.0.2
ensure==1.0.2
PyYAML
tqdm
glob2
types-PyYAML
joblib

## setup the project packages
-e .